+++ src/storage/catalog_class.c
@@ -693,7 +693,7 @@ catcls_find_oid_by_class_name (THREAD_ENTRY * thread_p, const char *name_p, OID 
   int error = NO_ERROR;
 
   error =
-    db_make_varchar (&key_val, DB_MAX_IDENTIFIER_LENGTH, (char *) name_p, strlen (name_p), LANG_SYS_CODESET,
+    db_make_varchar (&key_val, DB_MAX_IDENTIFIER_LENGTH, (char *) name_p, (int) strlen (name_p), LANG_SYS_CODESET,
 		     LANG_SYS_COLLATION);
   if (error != NO_ERROR)
     {
